ACM-17218-refactor-out AcmTableToolbar (#4602)
* ACM-17218-refactor-out AcmTableToolbar


* fix conflicting window.location between DetailsPage tests


* fix sonar issues


---------
