+++ frontend/src/ui-components/AcmTable/AcmTable.tsx
@@ -2,37 +2,19 @@
 
 import { css } from '@emotion/css'
 import {
-  Badge,
-  ButtonVariant,
   MenuToggle,
   PageSection,
   Pagination,
-  PaginationProps,
   PaginationVariant,
-  PerPageOptions,
   Skeleton,
   Toolbar,
-  ToolbarChip,
   ToolbarContent,
-  ToolbarFilter,
   ToolbarGroup,
   ToolbarItem,
-  Tooltip,
   TooltipPosition,
-  TooltipProps,
 } from '@patternfly/react-core'
-import {
-  Dropdown,
-  DropdownItem,
-  DropdownToggle,
-  DropdownToggleCheckbox,
-  Select,
-  SelectGroup,
-  SelectOption,
-  SelectOptionObject,
-  SelectVariant,
-} from '@patternfly/react-core/deprecated'
-import { EllipsisVIcon, ExportIcon, FilterIcon } from '@patternfly/react-icons'
+import { DropdownItem } from '@patternfly/react-core/deprecated'
+import { EllipsisVIcon } from '@patternfly/react-icons'
 import { css as cssPF } from '@patternfly/react-styles'
 import {
   ActionsColumn,
@@ -57,11 +39,9 @@ import {
   Tr,
 } from '@patternfly/react-table'
 import useResizeObserver from '@react-hook/resize-observer'
-import { debounce } from 'debounce'
 import Fuse from 'fuse.js'
 import get from 'get-value'
 import { mergeWith } from 'lodash'
-import { parse, ParsedQuery, stringify } from 'query-string'
 import {
   cloneElement,
   createContext,
@@ -73,411 +53,29 @@ import {
   useEffect,
   useLayoutEffect,
   useMemo,
+  useRef,
   useState,
 } from 'react'
-import { useLocation, useNavigate } from 'react-router-dom-v5-compat'
-import { HighlightSearchText } from '../../components/HighlightSearchText'
 import { useTranslation } from '../../lib/acm-i18next'
 import { usePaginationTitles } from '../../lib/paginationStrings'
 import { PluginContext } from '../../lib/PluginContext'
-import { FilterCounts, IRequestListView, IResultListView, IResultStatuses } from '../../lib/useAggregates'
-import { createDownloadFile, matchesFilterValue, parseLabel, returnCSVSafeString } from '../../resources/utils'
-import { IAlertContext } from '../AcmAlert/AcmAlert'
+import { createDownloadFile, returnCSVSafeString } from '../../resources/utils'
 import { AcmToastContext } from '../AcmAlert/AcmToast'
 import { AcmButton } from '../AcmButton/AcmButton'
-import { AcmDropdown, AcmDropdownItems } from '../AcmDropdown'
 import { AcmEmptyState } from '../AcmEmptyState/AcmEmptyState'
-import { AcmSearchInput, SearchConstraint, SearchOperator } from '../AcmSearchInput'
+import { SearchConstraint } from '../AcmSearchInput'
+import { setColumnValues, getColumnValues } from './localColumnStorage'
+import { AcmTableToolbar, applyFilters, ToolbarRef, useTableFilterSelections } from './AcmTableToolbar'
+import {
+  AcmTableProps,
+  AdvancedFilterSelection,
+  CommonPaginationPropsType,
+  CurrentFilters,
+  IAcmRowAction,
+  IAcmTableColumn,
+  ITableItem,
+} from './AcmTableTypes'
 import { AcmManageColumn } from './AcmManageColumn'
-import { filterLabelMargin, filterOption, filterOptionBadge } from './filterStyles'
-import { setLocalStorage, getLocalStorage, setColumnValues, getColumnValues } from './localColumnStorage'
-
-type SortFn<T> = (a: T, b: T) => number
-type CellFn<T> = (item: T, search: string) => ReactNode
-type SearchFn<T> = (item: T) => string | boolean | number | string[] | boolean[] | number[]
-
-// when a filter has more then this many options, give it its own dropdown
-const SPLIT_FILTER_THRESHOLD = 30
-// so we don't create 3000 elements, only create this many
-// with the assumption that if the user is looking for an option
-// they will use filter to find it
-const MAXIMUM_OPTIONS = 200
-
-/* istanbul ignore next */
-export interface IAcmTableColumn<T> {
-  /** the header of the column */
-  header: string
-
-  tooltip?: ReactNode
-
-  /** enables sort either on field name of using sort function */
-  sort?: SortFn<T> | string
-
-  /** if defined will enable search of the search field */
-  search?: SearchFn<T> | string
-
-  /** cell content, either on field name of using cell function */
-  cell: CellFn<T> | string
-
-  /** exported value as a string, supported export: CSV*/
-  exportContent?: CellFn<T>
-
-  disableExport?: boolean
-
-  transforms?: ITransform[]
-
-  cellTransforms?: ITransform[]
-
-  // Below this for column management
-  id?: string
-
-  order?: number
-
-  isDefault?: boolean
-  // If it is true, This column always the last one and isn't managed by column management filter
-  isActionCol?: boolean
-  // isFirstVisitChecked=true, When users visit at the first time, users can see these columns.
-  // unlike isDefualt columns, these columns can be controllable.
-  isFirstVisitChecked?: boolean
-
-  // Used to supply export information for Sub Rows. If true, the table will include column item within the CSV export
-  isSubRowExport?: boolean
-}
-
-/* istanbul ignore next */
-export interface IAcmRowAction<T> {
-  /** Action identifier */
-  id: string
-  /** Display a tooltip for this action */
-  tooltip?: string
-  /** Additional tooltip props forwarded to tooltip component */
-  tooltipProps?: Partial<TooltipProps>
-  /** Inject a separator horizontal rule immediately before an action */
-  addSeparator?: boolean
-  /** Display an action as being ariaDisabled */
-  isAriaDisabled?: boolean
-  /** Display an action as being disabled */
-  isDisabled?: boolean
-  /** Visible text for action */
-  title: string | React.ReactNode
-  /** Function for onClick() action */
-  click: (item: T) => void
-}
-
-/**
- * Type for table primary and secondary buttons.
- */
-export interface IAcmTableButtonAction {
-  id: string
-  title: string | React.ReactNode
-  click: () => void
-  isDisabled?: boolean | undefined
-  tooltip?: string | React.ReactNode
-  variant: ButtonVariant.primary | ButtonVariant.secondary
-}
-
-/**
- * Type for bulk actions on table items
- */
-export interface IAcmTableBulkAction<T> {
-  id: string
-  title: string | React.ReactNode
-  click: (items: T[]) => void
-  isDisabled?: ((items: T[]) => boolean) | boolean
-  tooltip?: string | React.ReactNode
-  variant: 'bulk-action'
-}
-
-/**
- * Type for separator line in action dropdown
- */
-export interface IAcmTableActionSeparator {
-  id: string
-  variant: 'action-separator'
-}
-
-/**
- * Type for table action dropdown options group
- */
-export interface IAcmTableActionGroup<T> {
-  id: string
-  title: string | React.ReactNode
-  actions: (IAcmTableBulkAction<T> | IAcmTableActionSeparator)[]
-  variant: 'action-group'
-}
-
-export type IAcmTableAction<T> = IAcmTableBulkAction<T> | IAcmTableActionSeparator | IAcmTableActionGroup<T>
-
-export interface ExportableIRow extends IRow {
-  // content from subrow to include in export document
-  exportSubRow?: {
-    header: string
-    exportContent: (item: any) => string
-  }[]
-}
-
-interface ITableItem<T> {
-  item: T
-  key: string
-  subRows?: ExportableIRow[]
-  [key: string]: unknown
-}
-
-type FilterOptionValueT = string
-type FilterSelection = FilterOptionValueT[]
-type TableFilterOption<FilterOptionValueT> = { label: ReactNode; value: FilterOptionValueT }
-
-type AdvancedFilterSelection = {
-  operator: SearchOperator
-  value: string
-}
-
-type CurrentFilters<S> = {
-  [filter: string]: S
-}
-
-type TableFilterBase<T, S> = {
-  /** unique identifier for the filter */
-  id: string
-  /** string displayed in the UI */
-  label: string
-  /** A required function that returns a boolean if the item is a match to the current filters */
-  tableFilterFn: (selection: S, item: T) => boolean
-}
-export interface ITableFilter<T> extends TableFilterBase<T, FilterSelection> {
-  /** Options is an array to define the exact filter options */
-  options: TableFilterOption<FilterOptionValueT>[]
-  showEmptyOptions?: boolean
-  supportsInequality?: boolean
-}
-interface IValidFilters<T> {
-  filter: ITableFilter<T>
-  options: { option: TableFilterOption<string>; count: number }[]
-}
-
-export interface ITableAdvancedFilter<T> extends TableFilterBase<T, AdvancedFilterSelection> {
-  availableOperators: SearchOperator[]
-}
-
-type TableFilterOptions = { option: TableFilterOption<string>; count: number }
-
-// render filter options with highlights for searched filter text
-// if option is a label like 'key=value', add a toggle button that toggles between = and !=
-function renderFilterSelectOption(
-  filterId: string,
-  option: TableFilterOptions,
-  supportsInequality?: boolean,
-  toggleEquality?: (filterId: string, option: TableFilterOptions) => void,
-  search?: string
-) {
-  const key = `${filterId}-${option.option.value}`
-  const handleInequality = () => {
-    toggleEquality?.(filterId, option)
-  }
-  return (
-    <SelectOption key={key} inputId={key} value={createFilterSelectOptionObject(filterId, option.option.value)}>
-      <div className={filterOption}>
-        <HighlightSearchText
-          text={(option.option.label as string) ?? '-'}
-          supportsInequality={supportsInequality}
-          toggleEquality={handleInequality}
-          searchText={search}
-        />
-        <Badge className={filterOptionBadge} key={key} isRead>
-          {option.count}
-        </Badge>
-      </div>
-    </SelectOption>
-  )
-}
-
-// filter options are retrieved from the url query and in the local storage
-// but they maybe old, so we make sure they're still valid by matching them up
-// with current filter options
-function getValidFilterSelections<T>(
-  filters: ITableFilter<T>[],
-  selections: CurrentFilters<FilterSelection> | ParsedQuery<string>
-) {
-  const validSelections: CurrentFilters<FilterSelection> = {}
-  let removedOptions = false
-  Object.keys(selections).forEach((key) => {
-    const filter = filters.find((filter) => filter.id === key)
-    if (filter) {
-      const filterValue = selections[key]
-      if (filterValue) {
-        // Normalize to array
-        let filterValues: (string | null)[] = []
-        if (Array.isArray(filterValue)) {
-          filterValues = filterValue
-        } else if (typeof filterValue === 'string') {
-          filterValues = [filterValue]
-        }
-
-        // Filter out invalid options
-        const supportsInequality = !!filter.supportsInequality
-        validSelections[key] = filterValues.filter((fv) => {
-          const inx = filter.options.findIndex(({ value }) => matchesFilterValue(supportsInequality, value, fv))
-          if (inx === -1) {
-            removedOptions = true
-            return false
-          }
-          return true
-        }) as string[]
-
-        // if none left
-        if (validSelections[key].length === 0) {
-          delete validSelections[key]
-        }
-      }
-    }
-  })
-  return { validSelections, removedOptions }
-}
-
-export function useTableFilterSelections<T>({ id, filters }: { id?: string; filters: ITableFilter<T>[] }) {
-  const tableFilterLocalStorageKey = id ? `acm-table-filter.${id}` : undefined
-
-  const { pathname, search } = useLocation()
-  const navigate = useNavigate()
-  const [queuedSearch, setQueuedSearch] = useState<string | null>(null)
-
-  const queryParams = useMemo(() => {
-    return parse(search, { arrayFormat: 'comma' })
-  }, [search])
-
-  const filteredQueryParams = useMemo(() => {
-    const filteredQueryParams: ParsedQuery<string> = {}
-    Object.keys(queryParams).forEach((key) => {
-      const filter = filters.find((filter) => filter.id === key)
-      if (!filter) {
-        filteredQueryParams[key] = queryParams[key]
-      }
-    })
-    return filteredQueryParams
-  }, [filters, queryParams])
-
-  const updateFilters = useCallback(
-    (newFilters: CurrentFilters<FilterSelection>, saveFilters: boolean = true) => {
-      const updatedParams = { ...filteredQueryParams, ...newFilters }
-      const updatedSearch = stringify(updatedParams, { arrayFormat: 'comma' })
-      const newSearch = updatedSearch ? `?${updatedSearch}` : ''
-
-      if (search !== newSearch) {
-        setQueuedSearch(newSearch)
-      }
-
-      if (saveFilters && tableFilterLocalStorageKey) {
-        setLocalStorage(tableFilterLocalStorageKey, newFilters)
-      }
-    },
-    [filteredQueryParams, search, tableFilterLocalStorageKey]
-  )
-
-  const filterSelections = useMemo(() => {
-    // Load filter selections from query params and validate
-    const { validSelections, removedOptions } = getValidFilterSelections(filters, queryParams)
-    if (Object.keys(validSelections).length) {
-      if (removedOptions) {
-        updateFilters(validSelections, false)
-      }
-      return validSelections
-    } else if (tableFilterLocalStorageKey) {
-      // if no query param filters, check local storage
-      const { validSelections } = getValidFilterSelections(filters, getLocalStorage(tableFilterLocalStorageKey, {}))
-      if (Object.keys(validSelections).length) {
-        updateFilters(validSelections, false)
-      }
-      return validSelections
-    }
-    return {}
-  }, [filters, tableFilterLocalStorageKey, queryParams, updateFilters])
-
-  const addFilterValue = useCallback(
-    (key: string, value: string) => {
-      const newFilter = { [key]: [value] }
-      const { validSelections } = getValidFilterSelections(filters, newFilter)
-      if (validSelections[key]?.length) {
-        const newFilters = { ...filterSelections, [key]: [...(filterSelections[key] || []), value] }
-        updateFilters(newFilters)
-      }
-    },
-    [filterSelections, filters, updateFilters]
-  )
-
-  const removeFilterValue = useCallback(
-    (key: string, value: string) => {
-      if (filterSelections[key]?.includes(value)) {
-        const newFilters = { ...filterSelections, [key]: filterSelections[key].filter((fv) => fv !== value) }
-        if (newFilters[key].length === 0) {
-          delete newFilters[key]
-        }
-        updateFilters(newFilters)
-      }
-    },
-    [filterSelections, updateFilters]
-  )
-
-  // for filters that are labels (ex: key=value), toggle the label between = and !=
-  const negateFilterValue = useCallback(
-    (key: string, value: string) => {
-      let newFilters
-      if (!filterSelections[key]?.includes(value)) {
-        const newFilter = { [key]: [value] }
-        const { validSelections } = getValidFilterSelections(filters, newFilter)
-        if (validSelections[key]?.length) {
-          newFilters = { ...filterSelections, [key]: [...(filterSelections[key] || []), value] }
-        }
-      } else {
-        newFilters = { ...filterSelections }
-      }
-      if (newFilters) {
-        const inx = newFilters[key].findIndex((fv) => fv === value)
-        const p = parseLabel(value)
-        const toggledValue = `${p.prefix}${p.oper === '=' ? '!=' : '='}${p.suffix}`
-        newFilters[key].splice(inx, 1, toggledValue)
-        updateFilters(newFilters)
-      }
-    },
-    [filterSelections, filters, updateFilters]
-  )
-
-  const removeFilter = useCallback(
-    (key: string) => {
-      if (filterSelections[key]) {
-        const newFilters = { ...filterSelections }
-        delete newFilters[key]
-        updateFilters(newFilters)
-      }
-    },
-    [filterSelections, updateFilters]
-  )
-
-  const clearFilters = useCallback(() => {
-    updateFilters({})
-  }, [updateFilters])
-
-  useEffect(() => {
-    if (queuedSearch !== null) {
-      navigate({ pathname, search: queuedSearch }, { replace: true })
-      setQueuedSearch(null)
-    }
-  }, [navigate, pathname, queuedSearch])
-
-  return { filterSelections, addFilterValue, removeFilterValue, negateFilterValue, removeFilter, clearFilters }
-}
-
-type FilterSelectOptionObject = SelectOptionObject & {
-  filterId: string
-  value: FilterOptionValueT
-}
-
-const createFilterSelectOptionObject = (filterId: string, value: FilterOptionValueT): FilterSelectOptionObject => ({
-  filterId,
-  value,
-  toString: () => value,
-  compareTo: (selectOption: FilterSelectOptionObject) =>
-    selectOption.filterId === filterId && selectOption.value === value,
-})
 
 const tableDivClass = css({
   display: 'table',
@@ -494,8 +92,6 @@ const tableClass = css({
     },
   },
 })
-
-export const SEARCH_DEBOUNCE_TIME = 500
 
 const DEFAULT_ITEMS_PER_PAGE = 10
 
@@ -528,22 +124,6 @@ export function AcmTablePaginationContextProvider(props: { children: ReactNode; 
   return <AcmTablePaginationContext.Provider value={paginationContext}>{children}</AcmTablePaginationContext.Provider>
 }
 
-const findFilterMatch = <T, S>(filter: string, filterArray: TableFilterBase<T, S>[]) =>
-  filterArray.find((filterItem) => filterItem.id === filter)
-
-const applyFilters = <T, S>(
-  items: ITableItem<T>[],
-  filterSelections: CurrentFilters<S>,
-  filterArray: TableFilterBase<T, S>[]
-): ITableItem<T>[] => {
-  const filterCategories = Object.keys(filterSelections)
-  return items.filter(({ item }) =>
-    filterCategories.every(
-      (filter: string) => findFilterMatch(filter, filterArray)?.tableFilterFn(filterSelections[filter], item) ?? true
-    )
-  )
-}
-
 function mergeProps(...props: any) {
   const firstProps = props[0]
   const restProps = props.slice(1)
@@ -570,52 +150,6 @@ function mergeProps(...props: any) {
   })
 }
 
-export type AcmTableProps<T> = {
-  items?: T[]
-  addSubRows?: (item: T) => IRow[] | undefined
-  initialSelectedItems?: T[]
-  disabledItems?: T[]
-  columns: IAcmTableColumn<T>[]
-  keyFn: (item: T) => string
-  customTableAction?: ReactNode
-  tableActionButtons?: IAcmTableButtonAction[]
-  tableActions?: IAcmTableAction<T>[]
-  rowActions?: IAcmRowAction<T>[]
-  rowActionResolver?: (item: T) => IAcmRowAction<T>[]
-  extraToolbarControls?: ReactNode
-  additionalToolbarItems?: ReactNode
-  emptyState: ReactNode
-  onSelect?: (items: T[]) => void
-  initialPage?: number
-  page?: number
-  setPage?: (page: number) => void
-  setRequestView?: (requestedView: IRequestListView) => void
-  resultView?: IResultListView
-  resultCounts?: IResultStatuses
-  fetchExport?: (requestedExport: IRequestListView) => Promise<IResultListView | undefined>
-  initialPerPage?: number
-  initialSearch?: string
-  search?: string
-  setSearch?: (search: string) => void
-  searchPlaceholder?: string
-  initialSort?: ISortBy | undefined
-  sort?: ISortBy | undefined
-  setSort?: (sort: ISortBy) => void
-  showToolbar?: boolean
-  gridBreakPoint?: TableGridBreakpoint
-  perPageOptions?: PerPageOptions[]
-  autoHidePagination?: boolean
-  noBorders?: boolean
-  fuseThreshold?: number
-  filters?: ITableFilter<T>[]
-  secondaryFilterIds?: string[]
-  advancedFilters?: ITableAdvancedFilter<T>[]
-  id?: string
-  showColumnManagement?: boolean
-  showExportButton?: boolean
-  exportFilePrefix?: string
-}
-
 export function AcmTable<T>(props: AcmTableProps<T>) {
   const {
     id,
@@ -626,16 +160,12 @@ export function AcmTable<T>(props: AcmTableProps<T>) {
     tableActions = [],
     rowActions = [],
     rowActionResolver,
-    customTableAction,
-    additionalToolbarItems,
     filters = [],
-    secondaryFilterIds,
     advancedFilters = [],
     gridBreakPoint,
     initialSelectedItems,
     onSelect: propsOnSelect,
     showColumnManagement,
-    showExportButton,
     exportFilePrefix,
     setRequestView,
     resultView,
@@ -643,14 +173,15 @@ export function AcmTable<T>(props: AcmTableProps<T>) {
     fetchExport,
   } = props
 
+  // a ref forwarded from toolbar to access its methods
+  const toolbarRef = useRef<ToolbarRef>(null)
   const defaultSort = {
     index: 0,
     direction: SortByDirection.asc,
   }
   const initialSort = props.initialSort || defaultSort
-  const initialSearch = props.initialSearch || ''
+  const initialSearch = props.initialSearch ?? ''
   const { isPreProcessed, loading, emptyResult } = resultView || {}
-  const { filterCounts } = resultCounts || {}
 
   const { t } = useTranslation()
   const toastContext = useContext(AcmToastContext)
@@ -677,24 +208,17 @@ export function AcmTable<T>(props: AcmTableProps<T>) {
   const [statePage, stateSetPage] = useState(props.initialPage || 1)
   const page = props.page || statePage
   const setPage = props.setPage || stateSetPage
-  const [stateSearch, stateSetSearch] = useState(initialSearch)
-  const search = props.search || stateSearch
-  const setSearch = props.setSearch || stateSetSearch
-  const searchPlaceholder = props.searchPlaceholder || t('Search')
   const [stateSort, stateSetSort] = useState<ISortBy | undefined>(initialSort)
   const sort = props.sort || stateSort
   const setSort = props.setSort || stateSetSort
   const [activeAdvancedFilters, setActiveAdvancedFilters] = useState<SearchConstraint[]>([])
-  const [pendingConstraints, setPendingConstraints] = useState<SearchConstraint[]>([
-    { operator: undefined, value: '', columnId: '' },
-  ])
 
   // State that is only stored in the component state
   const [selected, setSelected] = useState<{ [uid: string]: boolean }>({})
   const [disabled, setDisabled] = useState<{ [uid: string]: boolean }>({})
   const [preFilterSort, setPreFilterSort] = useState<ISortBy | undefined>(initialSort)
   const [expanded, setExpanded] = useState<{ [uid: string]: boolean }>({})
-  const [internalSearch, setInternalSearch] = useState(search)
+  const [internalSearch, setInternalSearch] = useState(props.search ?? initialSearch)
 
   // Dynamic gridBreakPoint
   const [breakpoint, setBreakpoint] = useState<TableGridBreakpoint>(TableGridBreakpoint.none)
@@ -704,7 +228,7 @@ export function AcmTable<T>(props: AcmTableProps<T>) {
   const outerDivRef = useCallback((elem: HTMLDivElement | null) => setOuterDiv(elem), [])
   const tableDivRef = useCallback((elem: HTMLDivElement | null) => setTableDiv(elem), [])
 
-  const { filterSelections, clearFilters } = useTableFilterSelections({ id, filters })
+  const { filterSelections } = useTableFilterSelections({ id, filters })
 
   //Column management
   const requiredColIds = useMemo(
@@ -816,25 +340,6 @@ export function AcmTable<T>(props: AcmTableProps<T>) {
       updateBreakpoint(width, tableWidth)
     }
   })
-
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  const setInternalSearchWithDebounce = useCallback(
-    process.env.NODE_ENV !== 'test'
-      ? debounce((search: string) => {
-          setInternalSearch(search)
-        }, SEARCH_DEBOUNCE_TIME)
-      : setInternalSearch,
-    [setInternalSearch]
-  )
-
-  useEffect(() => {
-    setInternalSearchWithDebounce(search)
-    return () => {
-      if ('clear' in setInternalSearchWithDebounce) {
-        setInternalSearchWithDebounce.clear()
-      }
-    }
-  }, [search, setInternalSearchWithDebounce])
 
   useEffect(() => {
     /* istanbul ignore else */
@@ -1027,85 +532,82 @@ export function AcmTable<T>(props: AcmTableProps<T>) {
     }
   }, [page, actualPage, setPage])
 
-  const exportTable = useCallback(
-    async (toastContext: IAlertContext) => {
-      toastContext.addAlert({
-        title: t('Generating data. Download may take a moment to start.'),
-        type: 'info',
-        autoClose: true,
-      })
+  const exportTable = useCallback(async () => {
+    toastContext.addAlert({
+      title: t('Generating data. Download may take a moment to start.'),
+      type: 'info',
+      autoClose: true,
+    })
 
-      const fileNamePrefix = exportFilePrefix ?? 'table-values'
-      const headerString: string[] = []
-      const csvExportCellArray: string[] = []
+    const fileNamePrefix = exportFilePrefix ?? 'table-values'
+    const headerString: string[] = []
+    const csvExportCellArray: string[] = []
 
-      columns.forEach(({ header, disableExport }) => {
-        if (header && !disableExport) {
-          headerString.push(header)
-        }
-      })
-      allTableItems[0].subRows?.[0]?.exportSubRow?.forEach(({ header }) => {
-        if (header) {
-          headerString.push(header)
-        }
-      })
-      csvExportCellArray.push(headerString.join(','))
-
-      // if table is pagenated from backend,
-      // we need to fetch all backend items to export
-      let exportItems = allTableItems
-      if (fetchExport) {
-        const fetchedItems = await fetchExport({
-          page: 1,
-          perPage: -1,
-          sortBy: undefined,
-        })
-        if (fetchedItems) {
-          exportItems = fetchedItems.items.map((item) => {
-            return {
-              item,
-            } as ITableItem<T>
-          })
-        }
+    columns.forEach(({ header, disableExport }) => {
+      if (header && !disableExport) {
+        headerString.push(header)
       }
+    })
+    allTableItems[0].subRows?.[0]?.exportSubRow?.forEach(({ header }) => {
+      if (header) {
+        headerString.push(header)
+      }
+    })
+    csvExportCellArray.push(headerString.join(','))
 
-      exportItems.forEach(({ item, subRows }) => {
-        let contentString: string[] = []
-        columns.forEach(({ header, exportContent, disableExport }) => {
-          if (header && !disableExport) {
-            // if callback and its output exists, add to array, else add "-"
-            const exportvalue = exportContent?.(item, '')
+    // if table is pagenated from backend,
+    // we need to fetch all backend items to export
+    let exportItems = allTableItems
+    if (fetchExport) {
+      const fetchedItems = await fetchExport({
+        page: 1,
+        perPage: -1,
+        sortBy: undefined,
+      })
+      if (fetchedItems) {
+        exportItems = fetchedItems.items.map((item) => {
+          return {
+            item,
+          } as ITableItem<T>
+        })
+      }
+    }
+
+    exportItems.forEach(({ item, subRows }) => {
+      let contentString: string[] = []
+      columns.forEach(({ header, exportContent, disableExport }) => {
+        if (header && !disableExport) {
+          // if callback and its output exists, add to array, else add "-"
+          const exportvalue = exportContent?.(item, '')
+          contentString.push(exportvalue ? returnCSVSafeString(exportvalue) : '-')
+        }
+      })
+      subRows?.forEach(({ exportSubRow }) => {
+        exportSubRow?.forEach(({ header, exportContent }) => {
+          if (header) {
+            const exportvalue = exportContent?.(item)
             contentString.push(exportvalue ? returnCSVSafeString(exportvalue) : '-')
           }
         })
-        subRows?.forEach(({ exportSubRow }) => {
-          exportSubRow?.forEach(({ header, exportContent }) => {
-            if (header) {
-              const exportvalue = exportContent?.(item)
-              contentString.push(exportvalue ? returnCSVSafeString(exportvalue) : '-')
-            }
-          })
-        })
-
-        contentString = [contentString.join(',')]
-        if (contentString[0]) {
-          csvExportCellArray.push(contentString[0])
-        }
       })
 
-      const exportString = csvExportCellArray.join('\n')
-      const fileName = `${fileNamePrefix}-${Date.now()}.csv`
+      contentString = [contentString.join(',')]
+      if (contentString[0]) {
+        csvExportCellArray.push(contentString[0])
+      }
+    })
 
-      createDownloadFile(fileName, exportString, 'text/csv')
+    const exportString = csvExportCellArray.join('\n')
+    const fileName = `${fileNamePrefix}-${Date.now()}.csv`
 
-      toastContext.addAlert({
-        title: t('Export successful'),
-        type: 'success',
-        autoClose: true,
-      })
-    },
-    [t, allTableItems, columns, exportFilePrefix, fetchExport]
-  )
+    createDownloadFile(fileName, exportString, 'text/csv')
+
+    toastContext.addAlert({
+      title: t('Export successful'),
+      type: 'success',
+      autoClose: true,
+    })
+  }, [toastContext, t, exportFilePrefix, columns, allTableItems, fetchExport])
 
   const paged = useMemo<ITableItem<T>[]>(() => {
     // if using a result view from backend, the items have already been sliced and diced
@@ -1198,51 +700,6 @@ export function AcmTable<T>(props: AcmTableProps<T>) {
           direction: sort.direction,
         }
       : sort
-
-  const clearSearch = useCallback(() => {
-    /* istanbul ignore if */
-    if (process.env.NODE_ENV !== 'test') {
-      ;(setInternalSearchWithDebounce as unknown as ReturnType<typeof debounce>).clear()
-    }
-    setSearch('')
-    setInternalSearch('')
-    setPage(1)
-    if (preFilterSort) {
-      setSort(preFilterSort)
-    }
-  }, [setSearch, setPage, preFilterSort, setInternalSearchWithDebounce, setSort])
-
-  const clearSearchAndFilters = useCallback(() => {
-    clearSearch()
-    clearFilters()
-    setActiveAdvancedFilters([])
-    setPendingConstraints([{ operator: undefined, value: '', columnId: '' }])
-  }, [clearSearch, clearFilters, setActiveAdvancedFilters, setPendingConstraints])
-
-  const updateSearch = useCallback(
-    (input: any) => {
-      // **Note: PatternFly change the fn signature
-      // From: (value: string, event: React.FormEvent<HTMLInputElement>) => void
-      // To: (_event: React.FormEvent<HTMLInputElement>, value: string) => void
-      // both cases need to be handled for backwards compatibility
-      const newSearch = typeof input === 'string' ? input : (input.target as HTMLInputElement).value
-      setSearch(newSearch)
-      setPage(1)
-      if (!newSearch) {
-        // clearing filtered state; restore previous sorting if applicable
-        if (preFilterSort) {
-          setSort(preFilterSort)
-        }
-      } else if (!search) {
-        // entering a filtered state; save sort setting use fuzzy match sort
-        setPreFilterSort(sort)
-        setSort({})
-      }
-    },
-    // setSort/setSearch/setPage can come from props, but setPreFilterSort is only from state and therefore
-    // guaranteed stable - not needed in dependency list
-    [search, sort, preFilterSort, setSort, setSearch, setPage]
-  )
 
   const updateSort = useCallback(
     (newSort: ISortBy) => {
@@ -1361,6 +818,25 @@ export function AcmTable<T>(props: AcmTableProps<T>) {
     [t]
   )
 
+  const renderColumnManagement = () => {
+    if (showColumnManagement) {
+      return (
+        <AcmManageColumn<T>
+          {...{
+            selectedColIds,
+            setSelectedColIds,
+            requiredColIds,
+            defaultColIds,
+            setColOrderIds,
+            colOrderIds,
+            tableId,
+          }}
+          allCols={columns.filter((col) => !col.isActionCol)}
+        />
+      )
+    }
+  }
+
   // Wrap provided action resolver
   const actionResolver = useMemo(
     () =>
@@ -1376,17 +852,15 @@ export function AcmTable<T>(props: AcmTableProps<T>) {
     [parseRowAction, resolveTableItem, rowActionResolver]
   )
 
-  const hasSearch = useMemo(() => columns.some((column) => column.search), [columns])
   const hasFilter = filters && filters.length > 0
   const hasItems = items && items.length > 0 && filtered
   const showToolbar =
     props.showToolbar !== false ? hasItems || emptyResult || (process.env.NODE_ENV !== 'test' && isLoading) : false
   const topToolbarStyle = items ? {} : { paddingBottom: 0 }
-  const [isExportMenuOpen, setIsExportMenuOpen] = useState(false)
 
   const translatedPaginationTitles = usePaginationTitles()
 
-  const commonPaginationProps: Partial<Omit<PaginationProps, 'ref'>> = {
+  const commonPaginationProps: CommonPaginationPropsType = {
     titles: translatedPaginationTitles,
     itemCount,
     perPage,
@@ -1433,157 +907,33 @@ export function AcmTable<T>(props: AcmTableProps<T>) {
         </Toolbar>
       )}
       {showToolbar && (
-        <Toolbar
-          clearFiltersButtonText={t('Clear all filters')}
-          clearAllFilters={clearSearchAndFilters}
-          collapseListedFiltersBreakpoint={'lg'}
-          inset={{ default: 'insetMd', xl: 'insetLg' }}
-        >
-          <ToolbarContent>
-            {hasSelectionColumn && (
-              <ToolbarItem>
-                <TableSelectionDropdown
-                  itemCount={itemCount}
-                  selectedCount={Object.keys(selected).length}
-                  perPage={perPage}
-                  onSelectNone={() => {
-                    const newSelected: { [uid: string]: boolean } = {}
-                    setSelected(newSelected)
-                  }}
-                  onSelectPage={() => {
-                    const newSelected: { [uid: string]: boolean } = {}
-                    for (const tableItem of paged) {
-                      newSelected[tableItem.key] = true
-                    }
-                    setSelected(newSelected)
-                    /* istanbul ignore next */
-                    if (propsOnSelect && items) {
-                      propsOnSelect(items.filter((item) => newSelected[keyFn(item)]))
-                    }
-                  }}
-                  onSelectAll={() => {
-                    const newSelected: { [uid: string]: boolean } = {}
-                    for (const tableItem of filtered) {
-                      if (!disabled[tableItem.key]) {
-                        newSelected[tableItem.key] = true
-                      }
-                    }
-                    setSelected(newSelected)
-                    /* istanbul ignore next */
-                    if (propsOnSelect && items) {
-                      propsOnSelect(items.filter((item) => newSelected[keyFn(item)]))
-                    }
-                  }}
-                />
-              </ToolbarItem>
-            )}
-            {(hasFilter || hasSearch) && (
-              <ToolbarGroup variant="filter-group">
-                {hasSearch && (
-                  <ToolbarItem variant="search-filter">
-                    <AcmSearchInput
-                      placeholder={searchPlaceholder}
-                      spellCheck={false}
-                      resultsCount={`${search === internalSearch ? filteredCount : '-'} / ${totalCount}`}
-                      style={{ flexGrow: 1 }}
-                      canAddConstraints
-                      useAdvancedSearchPopper={advancedFilters.length > 0}
-                      setActiveConstraints={setActiveAdvancedFilters}
-                      pendingConstraints={pendingConstraints}
-                      setPendingConstraints={setPendingConstraints}
-                      searchableColumns={advancedFilters.map((filter) => ({
-                        columnId: filter.id,
-                        columnDisplayName: filter.label,
-                        availableOperators: filter.availableOperators,
-                      }))}
-                      fuzzySearchValue={search}
-                      fuzzySearchOnChange={updateSearch}
-                      fuzzySearchOnClear={clearSearch}
-                    />
-                  </ToolbarItem>
-                )}
-                {hasFilter && (
-                  <TableColumnFilters
-                    id={id}
-                    filters={filters}
-                    secondaryFilterIds={secondaryFilterIds}
-                    filterCounts={filterCounts}
-                    items={items}
-                  />
-                )}
-              </ToolbarGroup>
-            )}
-            {props.tableActionButtons && props.tableActionButtons.length > 0 && (
-              <TableActionsButtons
-                actions={props.tableActionButtons}
-                hasSelections={Object.keys(selected).length > 0}
-              />
-            )}
-            {tableActions.length > 0 && (
-              <TableActions actions={tableActions} selections={selected} items={items} keyFn={keyFn} />
-            )}
-            {showColumnManagement && (
-              <AcmManageColumn<T>
-                {...{
-                  selectedColIds,
-                  setSelectedColIds,
-                  requiredColIds,
-                  defaultColIds,
-                  setColOrderIds,
-                  colOrderIds,
-                  tableId,
-                }}
-                allCols={columns.filter((col) => !col.isActionCol)}
-              />
-            )}
-            {customTableAction}
-            {showExportButton && (
-              <Tooltip content={t('Export all table data')}>
-                <ToolbarItem key={`export-toolbar-item`}>
-                  <Dropdown
-                    onSelect={(event) => {
-                      event?.stopPropagation()
-                      setIsExportMenuOpen(false)
-                    }}
-                    className="export-dropdownMenu"
-                    toggle={
-                      <DropdownToggle
-                        toggleIndicator={null}
-                        onToggle={(event, value) => {
-                          event.stopPropagation()
-                          setIsExportMenuOpen(value)
-                        }}
-                        aria-label="export-search-result"
-                        id="export-search-result"
-                      >
-                        <ExportIcon />
-                      </DropdownToggle>
-                    }
-                    isOpen={isExportMenuOpen}
-                    isPlain
-                    dropdownItems={[
-                      <DropdownItem key="export-csv" onClick={() => exportTable(toastContext)}>
-                        {t('Export all to CSV')}
-                      </DropdownItem>,
-                    ]}
-                    position={'left'}
-                  />
-                </ToolbarItem>
-              </Tooltip>
-            )}
-            {additionalToolbarItems}
-            {(!props.autoHidePagination || filtered.length > perPage) && (
-              <ToolbarItem variant="pagination">
-                <Pagination
-                  {...commonPaginationProps}
-                  aria-label={t('Pagination top')}
-                  isCompact
-                  perPageOptions={props.perPageOptions}
-                />
-              </ToolbarItem>
-            )}
-          </ToolbarContent>
-        </Toolbar>
+        <AcmTableToolbar
+          {...{
+            ...props,
+            hasFilter,
+            hasSelectionColumn,
+            commonPaginationProps,
+            sort,
+            setPage,
+            setSort,
+            preFilterSort,
+            setPreFilterSort,
+            selected,
+            setSelected,
+            disabled,
+            internalSearch,
+            setInternalSearch,
+            exportTable,
+            renderColumnManagement,
+            setActiveAdvancedFilters,
+            perPage,
+            paged,
+            filtered,
+            filteredCount,
+            totalCount,
+          }}
+          ref={toolbarRef}
+        />
       )}
       {!items || !rows || !filtered || !paged || (process.env.NODE_ENV !== 'test' && isLoading) ? (
         <PageSection variant="light" padding={{ default: 'noPadding' }}>
@@ -1752,7 +1102,7 @@ export function AcmTable<T>(props: AcmTableProps<T>) {
                 message={t('No results match the filter criteria. Clear filters to show results.')}
                 showSearchIcon={true}
                 action={
-                  <AcmButton variant="link" onClick={clearSearchAndFilters}>
+                  <AcmButton variant="link" onClick={toolbarRef?.current?.clearSearchAndFilters}>
                     {t('Clear all filters')}
                   </AcmButton>
                 }
@@ -1770,320 +1120,6 @@ export function AcmTable<T>(props: AcmTableProps<T>) {
         </Fragment>
       )}
     </Fragment>
-  )
-}
-
-function TableColumnFilters<T>(
-  props: Readonly<{
-    id?: string
-    filters: ITableFilter<T>[]
-    secondaryFilterIds?: string[]
-    filterCounts: FilterCounts | undefined
-    items?: T[]
-  }>
-) {
-  const [isOpen, setIsOpen] = useState([false])
-  const { id, filters, secondaryFilterIds, items, filterCounts } = props
-  const { filterSelections, addFilterValue, removeFilterValue, removeFilter, negateFilterValue } =
-    useTableFilterSelections({
-      id,
-      filters,
-    })
-  const { t } = useTranslation()
-
-  const onFilterSelect = useCallback(
-    (selection: FilterSelectOptionObject) => {
-      const { filterId, value } = selection
-      if (filterSelections[filterId]?.includes(value)) {
-        removeFilterValue(filterId, value)
-      } else {
-        addFilterValue(filterId, value)
-      }
-    },
-    [addFilterValue, filterSelections, removeFilterValue]
-  )
-
-  const onToggleEquality = useCallback(
-    (filterId: string, option: TableFilterOptions) => {
-      negateFilterValue(filterId, option.option.value)
-    },
-    [negateFilterValue]
-  )
-
-  const onDelete = useCallback(
-    (filter: string, chip: ToolbarChip) => {
-      removeFilterValue(filter, chip.key)
-    },
-    [removeFilterValue]
-  )
-
-  const onDeleteGroup = useCallback(
-    (filter: string) => {
-      removeFilter(filter)
-    },
-    [removeFilter]
-  )
-
-  const selections = useMemo(() => {
-    return Object.keys(filterSelections).reduce(
-      (acc: FilterSelectOptionObject[], filterId: string) =>
-        acc.concat(filterSelections[filterId].map((value) => createFilterSelectOptionObject(filterId, value))),
-      []
-    )
-  }, [filterSelections])
-
-  const filterSelectGroups = useMemo(() => {
-    const filterGroups = [
-      {
-        allFilters: [] as ITableFilter<T>[],
-        groupSelections: [] as FilterSelectOptionObject[],
-        validFilters: [] as IValidFilters<T>[],
-        allOptions: [] as TableFilterOptions[],
-      },
-    ]
-    for (const filter of filters) {
-      const supportsInequality = !!filter.supportsInequality
-      let options: TableFilterOptions[] = []
-      for (const option of filter.options) {
-        /* istanbul ignore next */
-        const count = filterCounts?.[filter.id]
-          ? filterCounts[filter.id][option.value]
-          : items?.filter((item) => filter.tableFilterFn([option.value], item)).length
-
-        // option is one of the options static filter options from our table (see derived table)
-        // we use that to find its matching current selection (from selections)
-        // we need to do special processing to match a selection like (key!=value)
-        //  because it won't directly match the original option (key=value)
-        const selectedOption = selections.find(
-          (selection) =>
-            selection.filterId === filter.id && matchesFilterValue(supportsInequality, option.value, selection.value)
-        )
-
-        // if the selection is a key!=value we can't use the original option (key=value)
-        // so we create a new option instead
-        const opt: TableFilterOption<string> =
-          supportsInequality && selectedOption
-            ? { label: selectedOption.value, value: selectedOption.value }
-            : { ...option }
-
-        /* istanbul ignore next */
-        if (
-          filter.showEmptyOptions ||
-          (count !== undefined && count > 0) ||
-          // if option is selected, it may be impacting results, so always show it even if options with 0 matches are being filtered
-          selectedOption
-        ) {
-          options.push({ option: opt, count: count ?? 0 })
-        }
-      }
-
-      // filter options can be spread out into multiple dropdowns if:
-      // 1 there's more then SPLIT_FILTER_THRESHOLD options or
-      // 2 the secondaryFilterId  (ex: secondaryFilterId = 'labels' will make it its own labels dropdown)
-      let group = filterGroups[0]
-      /* istanbul ignore else */
-      if (options.length) {
-        if (options.length > SPLIT_FILTER_THRESHOLD || secondaryFilterIds?.includes(filter.id)) {
-          options.sort((a, b) => {
-            return a?.option?.label?.toString().localeCompare(b?.option?.label?.toString() || '') || 0
-          })
-
-          filterGroups.push({
-            allFilters: [] as ITableFilter<T>[],
-            groupSelections: [] as FilterSelectOptionObject[],
-            validFilters: [] as IValidFilters<T>[],
-            allOptions: options,
-          })
-          // to avoid create lots of react components,
-          // just create a smaller set with the assumption that user
-          // won't be scrolling the entire list of 3000 clusters
-          // but will instead search for a cluster--at which point
-          // we will create react components for just that search
-          // in onFilterOptions
-          options = options.slice(0, MAXIMUM_OPTIONS)
-          group = filterGroups[filterGroups.length - 1]
-        }
-        group.validFilters.push({ filter, options })
-      }
-      group.allFilters.push(filter)
-    }
-
-    // if user has made selections and there are multiple filter dropdowns
-    // split the selections up by filter dropdown
-    filterGroups[0].groupSelections = selections
-    if (filterGroups.length > 1) {
-      let allSelections = [...selections]
-      filterGroups.forEach((group, inx) => {
-        if (inx !== 0) {
-          const remainingSelections = [] as FilterSelectOptionObject[]
-          filterGroups[inx].groupSelections = allSelections.filter((selected) => {
-            // there should only be one validFilter in extra filter dropdowns
-            // just for the type filter type (ex: cluster) in this dropdown
-            if (group.validFilters[0].filter.id !== selected.filterId) {
-              remainingSelections.push(selected)
-              return false
-            }
-            return true
-          })
-          allSelections = remainingSelections
-        }
-      })
-      filterGroups[0].groupSelections = allSelections
-    }
-
-    return filterGroups.map(({ allFilters, allOptions, groupSelections, validFilters }) => {
-      return {
-        groupFilters: allFilters,
-        groupOptions: allOptions,
-        groupSelections,
-        groupSelectionList: validFilters.map((filter) => {
-          return (
-            <SelectGroup key={filter.filter.id} label={filter.filter.label}>
-              {filter.options.map((option) => {
-                return renderFilterSelectOption(
-                  filter.filter.id,
-                  option,
-                  filter.filter.supportsInequality,
-                  onToggleEquality
-                )
-              })}
-            </SelectGroup>
-          )
-        }),
-      }
-    })
-  }, [filterCounts, filters, items, onToggleEquality, secondaryFilterIds, selections])
-
-  // used by filters with lots of options to filter the options
-  const onFilterOptions = useCallback(
-    (_: any, textInput: string, inx: number) => {
-      if (textInput !== '') {
-        const { id, supportsInequality } = filterSelectGroups[inx].groupFilters[0]
-        return filterSelectGroups[inx].groupOptions
-          .filter(({ option }) => {
-            return option?.value.toLowerCase().includes(textInput.toLowerCase())
-          })
-          .map((option) => {
-            return renderFilterSelectOption(id, option, supportsInequality, onToggleEquality, textInput.toLowerCase())
-          })
-      } else {
-        return filterSelectGroups[inx].groupSelectionList
-      }
-    },
-    [filterSelectGroups, onToggleEquality]
-  )
-
-  // create toolbar chips
-  const createChips = useCallback(
-    (current: ITableFilter<T>) => {
-      const currentCategorySelected = filterSelections[current.id] ?? []
-      // if options are made up of labels (key=value) just use the current selection values
-      if (current.supportsInequality) {
-        return currentCategorySelected.map((value) => {
-          return { key: value, node: value }
-        })
-      } else {
-        // else we need to get the correct label/value from derived table
-        return current.options
-          .filter((option: TableFilterOption<string>) => {
-            return currentCategorySelected.includes(option.value)
-          })
-          .map<ToolbarChip>((option: TableFilterOption<string>) => {
-            return { key: option.value, node: option.label }
-          })
-      }
-    },
-    [filterSelections]
-  )
-
-  return (
-    <ToolbarItem>
-      <div style={{ display: 'flex' }}>
-        {filterSelectGroups.map(({ groupFilters, groupSelections, groupSelectionList }, inx) => {
-          return groupFilters.reduce(
-            (acc, current) => (
-              <ToolbarFilter
-                key={'acm-table-filter-key'}
-                chips={createChips(current)}
-                deleteChip={(_category, chip) => {
-                  chip = chip as ToolbarChip
-                  onDelete(current.id, chip)
-                }}
-                deleteChipGroup={() => onDeleteGroup(current.id)}
-                categoryName={current.label}
-              >
-                {acc}
-              </ToolbarFilter>
-            ),
-            <Select
-              key={'acm-table-filter-select-key'}
-              variant={SelectVariant.checkbox}
-              aria-label={'acm-table-filter-select-key'}
-              onToggle={() => {
-                const arr = [...isOpen]
-                arr[inx] = !isOpen[inx]
-                setIsOpen(arr)
-              }}
-              onSelect={(
-                _event: React.MouseEvent<Element, MouseEvent> | React.ChangeEvent<Element>,
-                selection: SelectOptionObject
-              ) => onFilterSelect(selection as FilterSelectOptionObject)}
-              selections={groupSelections}
-              isOpen={isOpen[inx]}
-              isGrouped
-              placeholderText={
-                <div>
-                  <FilterIcon className={filterLabelMargin} />
-                  {inx === 0 ? t('Filter') : filterSelectGroups[inx].groupFilters[0].label}
-                </div>
-              }
-              noResultsFoundText={t('No results found')}
-              onFilter={(e, textInput) => onFilterOptions(e, textInput, inx)}
-              hasInlineFilter={inx !== 0}
-            >
-              {groupSelectionList}
-            </Select>
-          )
-        })}
-      </div>
-    </ToolbarItem>
-  )
-}
-
-function TableActions<T>(props: {
-  actions: IAcmTableAction<T>[]
-  selections: { [uid: string]: boolean }
-  items: T[] | undefined
-  keyFn: (item: T) => string
-}) {
-  const { actions, selections, items, keyFn } = props
-  /* istanbul ignore if */
-  if (actions.length === 0) return <Fragment />
-  return <TableActionsDropdown actions={actions} selections={selections} items={items} keyFn={keyFn} />
-}
-
-function TableActionsButtons(props: { actions: IAcmTableButtonAction[]; hasSelections?: boolean }) {
-  return (
-    <ToolbarGroup variant="button-group">
-      {props.actions.map((action) => {
-        /* istanbul ignore next */
-        const variant = props.hasSelections ? 'secondary' : action.variant
-        return (
-          <ToolbarItem key={`${action.id}-toolbar-item`}>
-            <AcmButton
-              id={action.id}
-              key={action.id}
-              onClick={action.click}
-              isDisabled={action.isDisabled}
-              tooltip={action.tooltip}
-              variant={variant}
-            >
-              {action.title}
-            </AcmButton>
-          </ToolbarItem>
-        )
-      })}
-    </ToolbarGroup>
   )
 }
 
@@ -2129,77 +1165,6 @@ function TableActionsButtons(props: { actions: IAcmTableButtonAction[]; hasSelec
  *
  * @returns A dropdown menu component for table bulk actions
  */
-
-function TableActionsDropdown<T>(props: {
-  actions: IAcmTableAction<T>[]
-  selections: { [uid: string]: boolean }
-  items: T[] | undefined
-  keyFn: (item: T) => string
-}) {
-  const { actions, selections = {}, items = [], keyFn } = props
-  const { t } = useTranslation()
-  const hasSelections = Object.keys(selections).length > 0
-
-  const dropdownItems = useMemo(() => {
-    function convertAcmTableActionsToAcmDropdownItems(actions: IAcmTableAction<T>[]): AcmDropdownItems[] {
-      return actions
-        .map((action, index) => {
-          if (action.variant === 'action-separator') {
-            return null
-          }
-          return {
-            id: action.id,
-            text: action.title,
-            separator: !!(index > 0 && actions[index - 1].variant === 'action-separator'),
-            ...(action.variant === 'action-group'
-              ? { flyoutMenu: convertAcmTableActionsToAcmDropdownItems(action.actions) }
-              : {
-                  tooltip: action.tooltip,
-                  isAriaDisabled:
-                    (typeof action.isDisabled === 'boolean' ? action.isDisabled : action.isDisabled?.(items)) ||
-                    !hasSelections,
-                }),
-          }
-        })
-        .filter((action) => action !== null)
-    }
-
-    return convertAcmTableActionsToAcmDropdownItems(actions)
-  }, [actions, items, hasSelections])
-
-  const handleSelect = useCallback(
-    (id: string) => {
-      // finds the action in both the top-level and the nested actions
-      const findAction = (actions: IAcmTableAction<T>[]): IAcmTableAction<T> | undefined => {
-        for (const action of actions) {
-          if (action.id === id) return action
-          if (action.variant === 'action-group') {
-            const nestedAction = findAction(action.actions)
-            if (nestedAction) return nestedAction
-          }
-        }
-        return undefined
-      }
-
-      const action = findAction(actions)
-      if (action && action.variant !== 'action-separator' && action.variant !== 'action-group') {
-        const selectedItems = items?.filter((item) => selections[keyFn(item)]) || []
-        action.click(selectedItems)
-      }
-    },
-    [actions, items, selections, keyFn]
-  )
-
-  return (
-    <AcmDropdown
-      id="table-actions-dropdown"
-      onSelect={handleSelect}
-      text={t('Actions')}
-      dropdownItems={dropdownItems}
-      isPrimary={hasSelections}
-    />
-  )
-}
 
 export function compareItems(path: string) {
   return (a: unknown, b: unknown) => {
@@ -2249,98 +1214,4 @@ export function compareNumbers(a: number | undefined | null, b: number | undefin
   if (a == undefined) return 1
   if (b == undefined) return -1
   return a < b ? -1 : a > b ? 1 : 0
-}
-
-export interface TableSelectionDropdownProps {
-  itemCount: number
-  selectedCount: number
-  perPage: number
-  onSelectNone: () => void
-  onSelectPage: () => void
-  onSelectAll: () => void
-}
-
-export function TableSelectionDropdown(props: TableSelectionDropdownProps) {
-  const [isOpen, setIsOpen] = useState(false)
-  const { itemCount, perPage, selectedCount, onSelectAll, onSelectNone, onSelectPage } = props
-  const [t] = useTranslation()
-  const onToggleCheckbox = useCallback(() => {
-    if (selectedCount > 0) onSelectNone()
-    else onSelectAll()
-  }, [selectedCount, onSelectNone, onSelectAll])
-
-  const toggleText = useMemo(() => {
-    return selectedCount > 0 ? t('{{count}} selected', { count: selectedCount }) : ''
-  }, [selectedCount, t])
-
-  const toggle = useMemo(() => {
-    return (
-      <DropdownToggle
-        splitButtonItems={[
-          <DropdownToggleCheckbox
-            id="select-all"
-            key="select-all"
-            aria-label={t('Select all')}
-            isChecked={selectedCount > 0}
-            onChange={onToggleCheckbox}
-          >
-            {toggleText}
-          </DropdownToggleCheckbox>,
-        ]}
-        onToggle={(_event, isOpen) => setIsOpen(isOpen)}
-      />
-    )
-  }, [t, selectedCount, onToggleCheckbox, toggleText])
-
-  const selectNoneDropdownItem = useMemo(() => {
-    return (
-      <DropdownItem
-        id="select-none"
-        key="select-none"
-        onClick={() => {
-          onSelectNone()
-          setIsOpen(false)
-        }}
-      >
-        {t('Select none')}
-      </DropdownItem>
-    )
-  }, [onSelectNone, t])
-
-  const selectPageDropdownItem = useMemo(() => {
-    return (
-      <DropdownItem
-        id="select-page"
-        key="select-page"
-        onClick={() => {
-          onSelectPage()
-          setIsOpen(false)
-        }}
-      >
-        {t('Select page ({{count}} items)', { count: Math.min(perPage, itemCount) })}
-      </DropdownItem>
-    )
-  }, [t, perPage, itemCount, onSelectPage])
-
-  const selectAllDropdownItem = useMemo(() => {
-    return (
-      <DropdownItem
-        id="select-all"
-        key="select-all"
-        onClick={() => {
-          onSelectAll()
-          setIsOpen(false)
-        }}
-      >
-        {t('Select all ({{count}} items)', { count: itemCount })}
-      </DropdownItem>
-    )
-  }, [t, itemCount, onSelectAll])
-
-  const dropdownItems = useMemo(
-    () => [selectNoneDropdownItem, selectPageDropdownItem, selectAllDropdownItem],
-    [selectNoneDropdownItem, selectPageDropdownItem, selectAllDropdownItem]
-  )
-
-  return <Dropdown isOpen={isOpen} toggle={toggle} dropdownItems={dropdownItems} />
 }
