+++ frontend/src/ui-components/index.ts
@@ -36,6 +36,8 @@ export * from './AcmSecondaryNav/AcmSecondaryNav'
 export * from './AcmSelect/AcmSelect'
 export * from './AcmSummaryList/AcmSummaryList'
 export * from './AcmTable/AcmTable'
+export * from './AcmTable/AcmTableToolbar'
+export * from './AcmTable/AcmTableTypes'
 export * from './AcmTextArea/AcmTextArea'
 export * from './AcmTextInput/AcmTextInput'
 export * from './AcmTile/AcmTile'
